refactor(ImageList): Create a new reusable components for displaying Images
