+++ web/src/components/Inputs/IconButton.tsx
@@ -23,7 +23,7 @@ export default function IconButton({
       " cursor-not-allowed text-slate-900/80 hover:text-slate-900/80";
   } else {
     buttonStyle +=
-      " text-slate-900 hover:text-indigo-700 active:scale-90 focus-visible:text-indigo-700 focus-visible:ring-4 focus-visible:ring-indigo-400/90 ring-inset";
+      " text-slate-900 hover:text-indigo-700 active:scale-90 focus-visible:text-indigo-700 focus-visible:ring-4 focus-visible:ring-indigo-700/90";
   }
 
   //#endregion
