+++ web/src/helpers/mutate.ts
@@ -1,0 +1,3 @@
+/*
+ * Functions to mutate data on the API
+ */
